feat(metrics): add overall metrics
