+++ package.json
@@ -14,6 +14,7 @@
     "react-router": "^5.0.0",
     "react-router-dom": "^5.0.0",
     "react-scripts": "3.0.1",
+    "recharts": "^1.6.2",
     "redux": "^4.0.0",
     "redux-react-hook": "^3.3.2"
   },
@@ -27,14 +28,15 @@
   "devDependencies": {
     "@types/fetch-mock": "^7.2.3",
     "@types/howler": "^2.1.0",
-    "@types/jest": "24.0.12",
+    "@types/jest": "^24.0.12",
     "@types/lodash": "^4.14.125",
     "@types/node": "^12.0.0",
-    "@types/react": "16.8.19",
-    "@types/react-dom": "16.8.4",
+    "@types/react": "^16.8.19",
+    "@types/react-dom": "^16.8.4",
     "@types/react-intl": "^2.3.17",
     "@types/react-router": "^5.0.0",
     "@types/react-router-dom": "^4.3.3",
+    "@types/recharts": "^1.1.16",
     "@types/redux-mock-store": "^1.0.1",
     "@typescript-eslint/eslint-plugin": "^1.6.0",
     "@typescript-eslint/parser": "^1.6.0",
