+++ src/components/metrics/user-metrics/index.ts
@@ -1,2 +1,1 @@
 export * from './metrics-view';
-export * from './user-metrics';
