+++ src/components/metrics/user-metrics/metrics-view.tsx
@@ -3,7 +3,7 @@ import React, { Suspense } from 'react';
 // @ts-ignore
 const LazyMetrics = React.lazy(() => import('./metrics'));
 
-export const MetricsView = () => {
+export const UserMetricsView = () => {
   return (
     <Suspense fallback="...">
       <LazyMetrics />
